requirements.txt
Django==5.2.8
google-generativeai==0.8.0
python-dotenv==1.1.1
